Added automatic IP detection
I suggest adding get_ip to automatically get IP seeing as it changes sometimes, I also suggest changing get_weathers location_req_url to get the IP everytime weather is called. Sorry if it is syntactically wrong, never really did anything in python, just saw that in the video it was static and decided this should be added.