voice_control.py
voice_control.py is a proof of concept that uses a free voice recognition API by Google.  This script only runs in Terminal for now.  "Mirror Mirror" is the audio trigger in order for the main part of the code to be run.  There would need to be a library of commands that it could recognize, as with Siri, Cortana, or OK Google.  The code needs to be threaded in order to interact with the smart mirror (like here: http://stackoverflow.com/questions/26719084/python-speech-recognition-in-tkinter). Requires speech_recognition and pyaudio libraries.