Prevent accidental check-ins of forecast.io tokens
Ignoring the file that contains the token means that developers can
safely test the application without having to manually reset the
line that contains their forecast.io token before making commits.

For users:

Instead of opening the source code and changing it directly, you
will need to create a new file in the root of the project called
`weather_api_token.txt`, and paste your token in there, then save it.